GPS-234: previewing a layer would mess up the zoom level
Issue: When previewing a layer, we sometimes got the issue where the layer preview would go to another zoom level. Also, sometimes, pressing the zoom buttons would change the zoom to a value that was far from what it was initially.

Cause: When using the `updateMapUrlSearchParams` function, we would send invalid values for the center and zoom, which caused an issue in the embed viewer and stopped the URL change within the embed. This meant that scrolling would technically function, but wouldn't change the URL, and consequently the store.

Fix: We set the center and zoom values to what they were before updating.